refactor(PageBuilder): replace ProductSection with FeaturedProducts and update import paths
