+++ components/shared/category/index.ts
@@ -1,0 +1,9 @@
+import FeaturedCategories from "./FeaturedCategories";
+import { FeaturedCategoryList } from "./FeaturedCategories";
+import { FeaturedCategoryThumb } from "./FeaturedCategories";
+
+export {
+    FeaturedCategories,
+    FeaturedCategoryList,
+    FeaturedCategoryThumb
+}
